use size_t type where appropriate

+++ bld/wl/h/spillio.h
@@ -37,8 +37,8 @@ typedef union {
 
 extern void             InitSpillFile( void );
 extern virt_mem_size    SpillAlloc( virt_mem_size );
-extern void             SpillNull( virt_mem_size, unsigned, unsigned  );
-extern void             SpillWrite( virt_mem_size, unsigned, void *, unsigned );
-extern void             SpillRead( virt_mem_size, unsigned, void *, unsigned );
+extern void             SpillNull( virt_mem_size, unsigned, size_t  );
+extern void             SpillWrite( virt_mem_size, unsigned, void *, size_t );
+extern void             SpillRead( virt_mem_size, unsigned, void *, size_t );
 extern void             CloseSpillFile( void );
 extern f_handle         OpenTempFile( char **name );
